Initial index.php
This is the initial upload for the index.php file which is used for obtaining the IP for the Hover Dyn-DNS updater. This is to be used on any remote system running a web server with PHP support in lieu of using one of the other sites out there that may, at times, get overloaded and go down or go to premium only service.